Move all class files to src/classes
